recipe generator/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
openai==0.28.0
flask-cors==4.0.0
python-decouple==3.8
